changed project structure, added .env.example, fixed .gitignore
