+++ .env.example
@@ -1,0 +1,1 @@
+REACT_APP_API_BASE_URL=your-api-base-url
