--- web/.env
@@ -1,1 +1,0 @@
-REACT_APP_API_BASE_URL=http://localhost:4000/api
